updated matrix library

+++ matrix/matrix_data.h
@@ -1,11 +1,13 @@
 #ifndef __MATRIX_DATA_H
 #define __MATRIX_DATA_H
 
+#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
+#include <string.h>
 
-typedef int matrix_data_t;
-typedef int matrix_size_t;
+typedef double matrix_data_t;
+typedef size_t matrix_size_t;
 
 typedef struct {
   matrix_size_t rows, collumns;
